Refactor edgedb generator into multiple files
