+++ src/core/edgedb/generator/query-builder.ts
@@ -1,0 +1,65 @@
+import { generateQueryBuilder as runQueryBuilderGenerator } from '@edgedb/generate/dist/edgeql-js.js';
+import { groupBy } from '@seedcompany/common';
+import { Directory, SyntaxKind } from 'ts-morph';
+import { customScalars } from './scalars';
+import { addCustomScalarImports, GeneratorParams } from './util';
+
+export async function generateQueryBuilder({
+  client,
+  root,
+  edgedbDir,
+}: GeneratorParams) {
+  const qbDir = edgedbDir.addDirectoryAtPath('generated-client');
+  await runQueryBuilderGenerator({
+    options: {
+      out: qbDir.getPath(),
+      updateIgnoreFile: false,
+      target: 'ts',
+      forceOverwrite: true,
+    },
+    client,
+    root: root.getPath(),
+  });
+  addJsExtensionDeepPathsOfEdgedbLibrary(qbDir);
+  changeCustomScalars(qbDir);
+  changeImplicitIDType(qbDir);
+}
+
+function addJsExtensionDeepPathsOfEdgedbLibrary(qbDir: Directory) {
+  for (const file of qbDir.addSourceFilesAtPaths('*')) {
+    const declarations = [
+      ...file.getImportDeclarations(),
+      ...file.getExportDeclarations(),
+    ].filter((d) => d.getModuleSpecifierValue()?.startsWith('edgedb/'));
+    for (const decl of declarations) {
+      decl.setModuleSpecifier(decl.getModuleSpecifierValue()! + '.js');
+    }
+  }
+}
+
+function changeCustomScalars(qbDir: Directory) {
+  // Change $uuid scalar type alias to use ID type instead of string
+  for (const scalars of groupBy(customScalars.values(), (s) => s.module)) {
+    const moduleFile = qbDir.addSourceFileAtPath(
+      `modules/${scalars[0]!.module}.ts`,
+    );
+    for (const scalar of scalars) {
+      const typeAlias = moduleFile.getTypeAliasOrThrow(`$${scalar.type}`);
+      typeAlias
+        .getFirstChildByKindOrThrow(SyntaxKind.TypeReference)
+        .getFirstChildByKindOrThrow(SyntaxKind.SyntaxList)
+        .getLastChildOrThrow()
+        .replaceWithText(scalar.ts);
+    }
+    addCustomScalarImports(moduleFile, scalars);
+  }
+}
+
+function changeImplicitIDType(qbDir: Directory) {
+  // Change implicit return shapes that are just the id to be ID type.
+  const typesystem = qbDir.addSourceFileAtPath(`typesystem.ts`);
+  addCustomScalarImports(typesystem, [customScalars.get('ID')!]);
+  typesystem.replaceWithText(
+    typesystem.getFullText().replaceAll('{ id: string }', '{ id: ID }'),
+  );
+}
